// src/utility/enum.ts
export enum NotifyEnums {
	SUCCESS = "success",
	ERROR = "error",
	WARNING = "warning",
	INFO = "info",
}

export enum StatusCodeEnums {
	UNAUTHORIZED = 401,
	FORBIDDEN = 403,
	NOT_FOUND = 404,
}

export enum ACTION_ENUM {
	EDIT = "EDIT",
	DELETE = "DELETE",
	ADD = "ADD",
	NAVIGATE = "NAVIGATE",
	MORE = "MORE",
	CUSTOM = "CUSTOM",
}

export enum TYPE_COMPONENT {
	LINK = "LINK",
	CHECKBOX = "CHECKBOX",
	BLOCK_CHECKBOX = "BLOCK-CHECKBOX",
	ACTIONS = "ACTIONS",
	TAG = "TAG",
	DATE = "DATE",
	CURRENCY = "CURRENCY",
	NUMBER = "NUMBER",
	INPUT = "INPUT",
	TEXT_AREA = "TEXT-AREA",
	INPUT_NUMBER = "INPUT-NUMBER",
	SELECT = "SELECT",
	DATE_PICKER = "DATE-PICKER",
	INPUT_PASSWORD = "INPUT-PASSWORD",
	TAG_LIST = "TAG-LIST",
	NUMBER_ORDER = "NUMBER-ORDER",
	RADIO_GROUP = "RADIO-GROUP",
	CUSTOM = "CUSTOM",
	NUMBER_PICKER = "NUMBER-PICKER",
	TEXT = "TEXT",
	LINK_COPY = "LINK-COPY",
	INPUT_NUMERIC = "INPUT-NUMERIC",
	TITLE_CELL = "TITLE-CELL",
	CURRENCY_FILTER = "CURRENCY-FILTER",
	CUSTOM_RENDER = "HARD-TEXT",
}

export enum STATUS {
	PENDING = "PENDING",
	ERROR = "ERROR",
	COMPLETED = "COMPLETED",
	APPROVED = "APPROVED",
	WAITING = "WAITING",
	DELETED = "DELETED",
	CREATED = "CREATED",
	REJECTED = "REJECTED",
}

export enum COLUMN_SIZING {
	AUTO_SIZE_ALL = "AUTO_SIZE_ALL",
	SIZE_TO_FIT = "SIZE_TO_FIT",
}

export enum PINNED {
	RIGHT = "right",
	LEFT = "left",
}

export enum STATUS_COMMAND {
	CONFIRMING = "CONFIRMING",
	COMPLETED = "COMPLETED",
	CANCELED = "CANCELED",
	WAITING = "WAITING",
	SENT = "SENT",
	CREATED = "CREATED",
	WAITING_PLAN = "WAITING_PLAN",
	REJECTED = "REJECTED",
	PENDING = "PENDING",
	REFUSED = "REFUSED",
	CANCEL = "CANCEL",
	WAITING_TRANSFER_STATION = "WAITING_TRANSFER_STATION",
}